% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%c\%}
\alias{\%c\%}
\title{Collapse a character vector to a character string}
\usage{
x \%c\% y
}
\arguments{
\item{x}{A character vector}

\item{y}{Character string, collapse parameter of paste0}
}
\value{
A character string
}
\details{
This function is a pipe version of paste0 with collapse parameter to specify
}
\examples{
c('N', 'F', 'K', 'A') \%c\% ' '
'All I wanna do' \%e\% '' \%c\% ''
}
