% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%count\%}
\alias{\%count\%}
\title{count pattern in strings the way of pipe operator}
\usage{
x \%count\% y
}
\arguments{
\item{x}{Pattern to count in string}

\item{y}{string where to count pattern}
}
\value{
A vector of integer
}
\examples{

fruit <- c("apple", "banana", "pear", "pineapple")
"a" \%count\% fruit
c("a", "b", "p", "p") \%count\% fruit
}
