% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%e\%}
\alias{\%e\%}
\title{Explode a character string}
\usage{
x \%e\% p
}
\arguments{
\item{x}{A character string}

\item{p}{A pattern}
}
\value{
a character vector
}
\description{
Explode a character string
}
\details{
This function is a pipe version of stringr::str_split for a character string, with simplify = TRUE
}
\examples{
'The little horse under the weather' \%e\% ' '
'The little horse under the weather' \%e\% ''
}
