% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%ee\%}
\alias{\%ee\%}
\title{Explode a character vector}
\usage{
x \%ee\% p
}
\arguments{
\item{x}{A character vector}

\item{p}{A pattern}
}
\value{
a tibble
}
\details{
This function is a pipe version of stringr::str_split for a character vector, with simplify = TRUE
}
\examples{
c('The little horse under the weather',
"He who gave courage",
"It was a little white horse",
'All behind, all behind',
"It was time of the little white horse",
"All behind him before") \%ee\% " |'|, "

}
