% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%g\%}
\alias{\%g\%}
\title{grepl in the way of pipe operator (case sensitive)}
\usage{
x \%g\% y
}
\arguments{
\item{x}{string where to search pattern with grepl}

\item{y}{pattern submitted to grepl}
}
\value{
A boolean
}
\details{
This functions is a pipe version of grepl
}
\examples{

'pig' \%g\% 'The pig is in the corn'
'Pig' \%g\% 'The pig is in the corn'
}
