% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%gic\%}
\alias{\%gic\%}
\title{grepl in the way of pipe operator (ignore case)}
\usage{
x \%gic\% y
}
\arguments{
\item{x}{string where to search pattern with grepl}

\item{y}{pattern submitted to grepl}
}
\value{
A boolean
}
\details{
This functions is a pipe version of grepl with ignore.case = TRUE
}
\examples{

'pig' \%gic\% 'The pig is in the cornfield'
'PIG' \%gic\% 'The PiG is in the cornfield'
}
