% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\% \%}
\alias{\% \%}
\alias{\%,\%}
\title{paste ', ' in the way of pipe operator}
\usage{
a \% \% b

a \%,\% b
}
\arguments{
\item{x}{A string (or coerced to)}

\item{y}{A string (or coerced to)}
}
\value{
The concatenation of \code{x} and \code{y} strings
}
\examples{
'Hello' \% \% 'world'
'Your pastas tastes like' \% \% '\%>\%'
'Hello' \%,\% 'world...'
'Your pastas tastes like ' \%+\% '\%>\%...' \%,\% 'or \%>>\%...'
}
