% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%out\%}
\alias{\%out\%}
\title{negation of in infix operator}
\usage{
a \%out\% b
}
\arguments{
\item{a}{Vector or NULL, values to be matched}

\item{b}{Vector or NULL, values to be matched}
}
\value{
a boolean
}
\description{
negation of in infix operator
}
\details{
This function comes from https://github.com/pierucci/rmngb/tree/master/rmngb
}
\examples{

'coca' \%out\% c('cola')
'coca' \%out\% c('coca', 'cola')
}
