% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%pad\%}
\alias{\%pad\%}
\title{Both side pad strings the way of pipe operator}
\usage{
string \%pad\% pad.width
}
\arguments{
\item{string}{Character vector or NULL}

\item{pad.width}{string or .digit where to indicate the pad and widht}
}
\value{
A character vector or NULL
}
\examples{

5 \%pad\% '.3'
5 \%rpad\% '0.3'
}
