% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%+\%}
\alias{\%+\%}
\title{paste0 in the way of pipe operator}
\usage{
a \%+\% b
}
\arguments{
\item{x}{A string (or coerced to)}

\item{y}{A string (or coerced to)}
}
\value{
The concatenation of \code{x} and \code{y} strings
}
\details{
This functions is a pipe version of paste0
}
\examples{
'Hello ' \%+\% 'world'
'Your pastas tastes like ' \%+\% '\%>\%'
'coco' \%+\% 'bolo'
}
