% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%re\%}
\alias{\%re\%}
\title{Replace matched patterns in a string}
\usage{
x \%re\% y.z
}
\arguments{
\item{x}{A character vector}

\item{y.z}{A vector of two string specifying pattern for match and replace}
}
\value{
A character vector
}
\details{
This function is a pipe version of stringr::str_replace
}
\examples{
'totoro12_13' \%re\% c('[0-9]+', 'to')

}
