% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%reall\%}
\alias{\%reall\%}
\title{Replace matched patterns in a string}
\usage{
x \%reall\% y.z
}
\arguments{
\item{x}{A character vector}

\item{y.z}{A vector of two string specifying pattern for match and replace}
}
\value{
A character vector
}
\description{
Replace matched patterns in a string
}
\details{
This function is a pipe version of stringr::str_replace_all
}
\examples{
'totoro12_13' \%reall\% c('[0-9]+', 'to')

}
