% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%rm\%}
\alias{\%rm\%}
\title{Remove matched patterns in a string}
\usage{
x \%rm\% y
}
\arguments{
\item{x}{A character vector}

\item{y}{Character string or pattern to remove}
}
\value{
A character vector
}
\details{
This function is a pipe version of stringr::str_remove
}
\examples{
'totoro' \%rm\% 'to'
}
