% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%rmall\%}
\alias{\%rmall\%}
\title{Remove matched patterns in a string}
\usage{
x \%rmall\% y
}
\arguments{
\item{x}{A character vector}

\item{y}{Character string or pattern to remove}
}
\value{
A character vector
}
\details{
This function is a pipe version of stringr::str_remove_all
}
\examples{
'totoro' \%rmall\% 'to'
}
