% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%rpad\%}
\alias{\%rpad\%}
\title{Right pad strings the way of pipe operator}
\usage{
string \%rpad\% pad.width
}
\arguments{
\item{string}{Character vector or NULL}

\item{pad.width}{string or .digit where to indicate the pad and width}
}
\value{
A character vector or NULL
}
\description{
Right pad strings the way of pipe operator
}
\details{
This functions is a pipe version of stringr::pad
}
\examples{

5 \%rpad\% '.2'
'é' \%rpad\% 'è.5'
}
