% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{\%s\%}
\alias{\%s\%}
\title{substring in the way of pipe operator}
\usage{
x \%s\% yz
}
\arguments{
\item{x}{Vector or NULL, values to be substringed}

\item{y.z}{A character or 0.digit to call substr}
}
\value{
A string or NULL
}
\description{
substring in the way of pipe operator
}
\details{
This functions is a pipe version of substr
}
\examples{

'NFKA008' \%s\% '1.4'
'NFKA008' \%s\% .4
'where is' \% \% ('the pig is in the cornfield' \%s\% '1.7') \%+\% '?'
'NFKA008' \%s\% 5.7
}
