% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{run_transcoder}
\alias{run_transcoder}
\title{Run transcoder shiny app}
\usage{
run_transcoder(
  launch.browser = getOption("shiny.launch.browser", interactive())
)
}
\description{
Run transcoder shiny app
}
