% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{tolower_names}
\alias{tolower_names}
\title{Make names of a tibble tolower}
\usage{
tolower_names(d)
}
\arguments{
\item{d}{tibble}
}
\value{
A tibble
}
\examples{
library(magrittr)
iris \%>\% tolower_names \%>\% head
}
