% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pipe_all.R
\name{toupper_names}
\alias{toupper_names}
\title{Make names of a tibble toupper}
\usage{
toupper_names(d)
}
\arguments{
\item{d}{tibble}
}
\value{
A tibble
}
\description{
Make names of a tibble toupper
}
\examples{
library(magrittr)
iris \%>\% toupper_names \%>\% head
}
